refactor: update main entry point to index.js and adjust root directory in tsconfig.json; remove unused scripts
