+++ package.json
@@ -2,11 +2,11 @@
   "name": "eradrin-bot",
   "version": "1.1.0",
   "description": "",
-  "main": "dist/bot.js",
+  "main": "index.js",
   "scripts": {
     "build": "tsc",
-    "start": "node dist/bot.js",
-    "dev": "concurrently \"tsc -w\" \"nodemon dist/bot.js\"",
+    "start": "node dist/index.js",
+    "dev": "concurrently \"tsc -w\" \"nodemon dist/index.js\"",
     "deploy:commands": "tsc && node dist/deploy-commands.js",
     "safe-start": "tsx scripts/safe-start.ts",
     "safe-start:js": "tsc && node scripts/safe-start.js",
