+++ src/scripts/clear-knowledge.ts
