+++ src/scripts/fetchAllCharacters.ts
