+++ src/scripts/processCharacters.ts
