+++ src/scripts/safe-start.ts
