+++ src/scripts/verify-knowledge.ts
